==> SVM/15.SVM_find_union-work.py <==
import cv2
import random
from scipy.io import *
import pandas as pd
import numpy as np
import tensorflow as tf

pred = pd.read_csv('../dataset/SVM_pred.csv', header=None)
preds = pred.values

print(pred.shape)  # (6, 5)

row = pred.shape[0]  # 6
col = pred.shape[1]  # 5

pred = np.transpose(pred)


# 给定一个二维（x，y）坐标，返回一个像素坐标的上下左右四个（x，y）坐标
def return_4loc(i, j):
    # up
    i_up = i - 1
    j_up = j
    # down
    i_down = i + 1
    j_down = j
    # left
    i_left = i
    j_left = j - 1
    # right
    i_right = i
    j_right = j + 1
    # 超出边界的像素返回四个FALSE
    if i_up < 0 or i_down > (row - 1) or j_left < 0 or j_right > (col - 1):
        return False, False, False, False
    else:
        # 否则返回四个坐标
        # print([i_up, j_up], [i_down, j_down], [i_left, j_left], [i_right, j_right])
        return [i_up, j_up], [i_down, j_down], [i_left, j_left], [i_right, j_right]


# 记录当前连通区域的一维坐标集合
unionloc_set = set()
# 记录当前连通区域边界像素的一维坐标集合
unionbound_set = set()
unionbound_stack = []

# 记录当前中心位置和对应联通区域大小的字典
union_size = dict()

# 判断是否边界触及边界
flag = True

# 记录已经处理的像素
stack = []


# 连通区域递归函数，传入中心目标像素的二维坐标（x，y）
def mani(i, j):
    # 记录当前中心像素的判定结果
    current_pred = pred[i][j]
    # 二维坐标换算一维坐标值备用
    loc = i * col + j

    # 获取当前元素上下左右四个二维坐标
    up, down, left, right = return_4loc(i, j)
    # 在当前像素处于画面边缘时，跳过
    if not up or not down or not left or not right:
        global flag
        flag = False
        return
    # 将获取到的上下左右四个二维坐标详细记录
    # up坐标
    i_up = up[0]
    j_up = up[1]
    # down坐标
    i_down = down[0]
    j_down = down[1]
    # left坐标
    i_left = left[0]
    j_left = left[1]
    # right坐标
    i_right = right[0]
    j_right = right[1]
    # 统计上下左右四个的一维坐标
    up_loc = i_up * col + j_up
    down_loc = i_down * col + j_down
    left_loc = i_left * col + j_left
    right_loc = i_right * col + j_right
    # 如果当前像素，对应的上下左右四个像素有一个不在gt标记范围内，就暂时跳过，这说明该像素在边缘
    if pred[i_up][j_up] == 0 or pred[i_down][j_down] == 0 or pred[i_left][j_left] == 0 or pred[i_right][
        j_right] == 0:
        flag = False
        return
    # 否则既然属于标记范畴，开始判定连通问题

    # 将当前中心元素率先放入连通集合
    unionloc_set.add(loc)

    # 再将中心区域上下左右四个像素放入边界集合，入栈
    if up_loc not in unionloc_set:
        # unionbound_set.add(up_loc)
        unionbound_stack.append(up_loc)
    if down_loc not in unionloc_set:
        # unionbound_set.add(down_loc)
        unionbound_stack.append(down_loc)
    if left_loc not in unionloc_set:
        # unionbound_set.add(left_loc)
        unionbound_stack.append(left_loc)
    if right_loc not in unionloc_set:
        # unionbound_set.add(right_loc)
        unionbound_stack.append(right_loc)

    # 限定联通面积阈值
    # if len(unionloc_set) > 20:
    #     return

    while len(unionbound_stack) != 0:
        # 弹出栈尾最后一个元素
        bound = unionbound_stack.pop(-1)
        i_bound = bound // col
        j_bound = bound % col
        if pred[i_bound][j_bound] == current_pred:
            mani(i_bound, j_bound)


for i in range(row):
    for j in range(col):
        print((i, j))
        # 记录当前连通区域的一维坐标集合
        unionloc_set = set()
        # 记录当前连通区域边界像素的一维坐标集合
        # unionbound_set = set()
        unionbound_stack = []
        # 确定联通区域是否触及边界
        flag = True
        cur_loc = i * row + j
        if pred[i][j] != 0:
            mani(i, j)
            if len(unionloc_set) < 20 and len(unionloc_set) != 0 and flag:
                union_size[(i, j)] = len(unionloc_set)
                for pix in unionloc_set:
                    i_pix = pix // col
                    j_pix = pix % col
                    pred[i_pix][j_pix] = 0

print(union_size)
# {(1, 1): 9, (1, 9): 2, (3, 15): 16, (6, 21): 7, (7, 10): 3, (9, 5): 1}
